Allowing a shape to be selector scope
A shape can now be used as a selector scope allowing for selecting and
comparing multiple top-level attributes of shapes.

This commit also updates AttributeValue to be ready to expose as a
public interface if needed for more advanced selector message formatting
functionality.
