+++ smithy-model/src/test/resources/software/amazon/smithy/model/selector/nested-traits.smithy
@@ -111,3 +111,7 @@ list ListyTraitMember2 {
 structure ListyTraitStruct {
     foo: String
 }
+
+@trait
+@recursiveTrait("Wow!")
+string recursiveTrait
